requirements.txt
# Core dependencies
requests>=2.32.0
psycopg2-binary>=2.9.11
SQLAlchemy>=2.0.44
geopandas>=1.1.1
pandas>=2.3.3
shapely>=2.1.2
pyproj>=3.7.2
Fiona>=1.10.1
geoalchemy2>=0.18.0
python-dotenv>=1.1.1
backoff>=2.2.1
topojson>=1.10
brotli>=1.1.0

# Standard library dependencies (included with Python)
# - time
# - datetime
# - concurrent.futures
# - json
# - io
# - os
# - sys
# - subprocess
# - argparse
# - pathlib
# - configparser
# - logging
# - unittest.mock

# Testing dependencies
pytest>=8.4.2
pytest-mock>=3.15.1
pytest-cov>=4.0.0
pytest-html>=3.1.0
responses>=0.25.8

# Development dependencies
# - typing_extensions (for type hints)
